[extras][script] emscripten: log to JS console
