+++ extras/fire_script_js/index.html
@@ -7,15 +7,13 @@
     <style>
         body { background-color: black; margin: 0; height: 100vh; }
         #border { height: 100%; padding: 2px; box-sizing: border-box; }
-        #terminal { height: 80%; }
-        #terminal_log { height: 20%; }
+        #terminal { height: 100%; }
     </style>
     <script src="fire.js"></script>
 </head>
 <body>
 <div id="border">
     <div id="terminal"></div>
-    <div id="terminal_log"></div>
 </div>
 <script src="bundle.js" defer></script>
 </body>
