+++ extras/fire_script_js/main.js
@@ -3,7 +3,6 @@ import { FitAddon } from 'xterm-addon-fit';
 import { WebLinksAddon } from 'xterm-addon-web-links';
 
 const termContainer = document.getElementById('terminal');
-const termLogContainer = document.getElementById('terminal_log');
 
 const term = new Terminal({
     convertEol: true,
@@ -14,42 +13,25 @@ const term = new Terminal({
 });
 term.open(termContainer);
 
-const termLog = new Terminal({
-    disableStdin: true,
-    convertEol: true,
-    rightClickSelectsWord: false,
-    theme: {
-        background: '#111',
-        cursor: '#111',
-    }
-});
-termLog.open(termLogContainer);
-
+// fit terminal size to window size
 const termFit = new FitAddon();
 term.loadAddon(termFit);
 termFit.fit();
-
-const termLogFit = new FitAddon();
-termLog.loadAddon(termLogFit);
-termLogFit.fit();
-
 window.onresize = function () {
     termFit.fit();
-    termLogFit.fit();
 };
 
 term.loadAddon(new WebLinksAddon());
 
-termLog.write("Loading...\n");
+term.write("Loading...\n");
 
 fire_script().then(Module => {
-    termLog.clear();
-    termLog.write("\x1b[J");  // clear screen from cursor down
+    term.clear();
+    term.write("\x1b[J");  // clear screen from cursor down
     const url_args = new URLSearchParams(window.location.search);
     const debug = Boolean(parseInt(url_args.get('debug')));
     const prog = new Module.FireScript(debug);
     prog.set_term_out_cb(data => term.write(data));
-    prog.set_term_err_cb(data => termLog.write(data));
     prog.set_quit_cb(function() {
         // try to actually quit, it may work in some circumstances
         window.close();
